* rename setup to settings * fix db update
