+++ console/src/app/app.module.ts
@@ -14,7 +14,7 @@ import { AgentsService } from './agents/agents.service';
 import { ProjectsService } from './projects/projects.service';
 import { CalendarService } from './calendar/calendar.service';
 import { CoreModule } from './core/core.module';
-import { SettingsService } from './core/setup/setup.service';
+import { SettingsService } from './core/setup/settings.service';
 import { IsSetUpGuard } from './core/setup/issetup.guard';
 import { VaultService } from '@shared/vault.service';
